+++ src/robot.impl.cc
@@ -845,7 +845,7 @@ namespace hpp
 	kppPolyhedron->CkcdPolyhedron::addPoint((kcdReal)x, (kcdReal)y,
 						(kcdReal)z, rank);
 
-	return rank;
+	return (Short)rank;
       }
 
 
@@ -876,7 +876,7 @@ namespace hpp
 
 	kppPolyhedron->addTriangle(pt1, pt2, pt3, rank);
 
-	return rank;
+	return (Short)rank;
       }
 
 
@@ -973,7 +973,7 @@ namespace hpp
 	  CkppDeviceComponentShPtr hppRobot = planner_->robotIthProblem(hppProblemId);
 
 
-	  outDeviceDim = hppRobot->countDofs ();
+	  outDeviceDim = (UShort)(hppRobot->countDofs ());
 	}
 	else{
 	  hppDout (error, "wrong robot Id");
